+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## NEXT
+* Update `KeyAttestationJwt` to [EUDI Wallet TS3](https://github.com/eu-digital-identity-wallet/eudi-doc-standards-and-technical-specifications/blob/main/docs/technical-specifications/ts3-wallet-unit-attestation.md) from 2026-03-15
+
 ## 3.21.0 / Supreme 0.13.0
 * Drop Apple X64 targets
 * Move `InstantLongSerializer` from indispensable-josef/indispensable-cosef to indispensable